fix: simulation 학과 선택 모달 오류 수정 (#61)
* fix: 시뮬레이션 불필요한 상태 정리

* fix: 시뮬레이션 진행 판별 오류 해결

* fix: 시뮬레이션 확인 함수 분리

* fix: 종료 후 시작 모달 뜨는 오류 수정

* fix: 시뮬레이션 테스트 코드 수정

* refactor: 불필요한 주석 제거

* refactor: 필요 없는 코드 제거
